Move changes_environ to tests/utils.py
