+++ python/tests/util.py
@@ -1,0 +1,31 @@
+#
+#   Copyright 2024 Hopsworks AB
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+
+import os
+from functools import wraps
+
+
+def changes_environ(f):
+    @wraps(f)
+    def g(*args, **kwds):
+        old_environ = os.environ.copy()
+        try:
+            return f(*args, **kwds)
+        finally:
+            os.environ.clear()
+            os.environ.update(old_environ)
+
+    return g
